+++ lib/src/channel.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: avoid_print
-
 import 'package:flutter_logger_plus/src/color_sequence.dart';
 import 'package:flutter_logger_plus/src/date.dart';
 
@@ -8,98 +6,66 @@ import 'config.dart';
 final channel = _Channel();
 
 class _Channel {
-  List<String> logs = [];
-
-  void error(String string, {showLog = false}) {
+  void error(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.red}ERROR: ${color.pink}$hms${color.red} | $string ${color.end}";
     print(string);
   }
 
-  void success(String string, {showLog = false}) {
+  void success(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.green}SUCCESS: ${color.pink}$hms${color.green} | $string ${color.end}";
     print(string);
   }
 
-  void info(String string, {showLog = false}) {
+  void info(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.yellow}INFO: ${color.pink}$hms${color.yellow} | $string ${color.end}";
     print(string);
   }
 
-  void log(String string, {showLog = false}) {
+  void white(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.white}PRINT: ${color.boldBlue}$hms${color.white} | $string ${color.end}";
     print(string);
   }
 
-  void blue(String string, {showLog = false}) {
+  void blue(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.blue}PRINT: ${color.pink}$hms${color.blue} | $string ${color.end}";
     print(string);
   }
 
-  void pink(String string, {showLog = false}) {
+  void pink(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.pink}PRINT: ${color.yellow}$hms${color.pink} | $string ${color.end}";
     print(string);
   }
 
-  void cyan(String string, {showLog = false}) {
+  void cyan(String string, bool showMilliSeconds) {
     if (isInProduction) return;
 
-    if (showLog) logs.add(string);
-
-    String hms = date.getHMS();
+    String hms = date.getHMS(showMilliSeconds);
     string =
         "${color.cyan}PRINT: ${color.boldBlue}$hms${color.cyan} | $string ${color.end}";
-    log(string);
-  }
-
-  void showLogs() {
-    if (logs.isEmpty) {
-      error("NO LOGS FOUND");
-      return;
-    }
-    for (String log in logs) {
-      print(log);
-    }
-  }
-
-  void clearLogs() {
-    logs.clear();
-    success("LOGS CLEARED!");
-    return;
+    print(string);
   }
 }