+++ lib/src/date.dart
@@ -2,11 +2,19 @@ final date = _Date();
 
 @override
 class _Date {
-  String getHMS({bool semicolonFormat = true}) {
+  String getHMS(bool showMilliSeconds, {bool semicolonFormat = true}) {
     DateTime now = DateTime.now();
+    
     int hour = now.hour;
     int minute = now.minute;
     int second = now.second;
+
+    if (!showMilliSeconds) {
+      return semicolonFormat
+          ? "$hour:$minute:$second"
+          : "$hour/$minute/$second";
+    }
+
     int millisecond = now.millisecond;
     return semicolonFormat
         ? "$hour:$minute:$second:$millisecond"