+++ lib/src/flutter_logger_plus.dart
@@ -4,39 +4,31 @@ final logger = _FlutterLoggerPlus();
 
 @override
 class _FlutterLoggerPlus {
-  void error(dynamic string) {
-    channel.error(string.toString());
+  void error(dynamic string, {bool showMilliSeconds = false}) {
+    channel.error(string.toString(), showMilliSeconds);
   }
 
-  void log(dynamic string) {
-    channel.log(string.toString());
+  void white(dynamic string, {bool showMilliSeconds = false}) {
+    channel.white(string.toString(), showMilliSeconds);
   }
 
-  void success(dynamic string) {
-    channel.success(string.toString());
+  void success(dynamic string, {bool showMilliSeconds = false}) {
+    channel.success(string.toString(), showMilliSeconds);
   }
 
-  void info(dynamic string) {
-    channel.info(string.toString());
+  void info(dynamic string, {bool showMilliSeconds = false}) {
+    channel.info(string.toString(), showMilliSeconds);
   }
 
-  void blue(dynamic string) {
-    channel.blue(string.toString());
+  void blue(dynamic string, {bool showMilliSeconds = false}) {
+    channel.blue(string.toString(), showMilliSeconds);
   }
 
-  void pink(dynamic string) {
-    channel.pink(string.toString());
+  void pink(dynamic string, {bool showMilliSeconds = false}) {
+    channel.pink(string.toString(), showMilliSeconds);
   }
 
-  void cyan(dynamic string) {
-    channel.cyan(string.toString());
-  }
-
-  void clearLogs() {
-    channel.clearLogs();
-  }
-
-  void showLogs() {
-    channel.showLogs();
+  void cyan(dynamic string, {bool showMilliSeconds = false}) {
+    channel.cyan(string.toString(), showMilliSeconds);
   }
 }