+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=D:\Software\src\flutter"
-export "FLUTTER_APPLICATION_PATH=D:\Flutter\flutter_logger_plus"
+export "FLUTTER_APPLICATION_PATH=D:\Projects\Flutter\flutter_logger_plus"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=0.0.1"
-export "FLUTTER_BUILD_NUMBER=0.0.1"
+export "FLUTTER_BUILD_NAME=4.0.3"
+export "FLUTTER_BUILD_NUMBER=4.0.3"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"